+++ src/app/socials/page.tsx
@@ -2,6 +2,6 @@
 
 import { LinksPageView } from '@/components/links/LinksPageView'
 
-export default function LinksPage() {
+export default function SocialsPage() {
 	return <LinksPageView />
 }